All crud

Related work items: #231, #232, #235, #238, #269, #272, #274, #275, #276, #300, #307, #309, #310, #311, #316, #317, #318, #321, #325, #331, #337, #338, #349, #351, #355, #359

+++ DijiWalk.Repositories/PlayerRepository.cs
@@ -5,8 +5,15 @@
 //-----------------------------------------------------------------------
 namespace DijiWalk.Repositories
 {
+    using System;
     using System.Collections.Generic;
     using System.Threading.Tasks;
+    using DijiWalk.Business;
+    using DijiWalk.Business.Contracts;
+    using DijiWalk.Common.Contracts;
+    using DijiWalk.Common.Encryption;
+    using DijiWalk.Common.FileExtension;
+    using DijiWalk.Common.Response;
     using DijiWalk.Entities;
     using DijiWalk.EntitiesContext;
     using DijiWalk.Repositories.Contracts;
@@ -19,33 +26,92 @@ namespace DijiWalk.Repositories
     {
         private readonly SmartCityContext _context;
 
+        private readonly ITeamBusiness _teamBusiness;
+
+        private readonly ICryption _cryption;
+
+        private readonly IPlayerBusiness _playerBusiness;
+
+        private readonly IImageBusiness _imageBusiness;
+
         /// <summary>
         /// Parameter that serve to connect to the database
         /// </summary>
-        public PlayerRepository(SmartCityContext context)
+        public PlayerRepository(SmartCityContext context, ITeamBusiness teamBusiness, ICryption cryption, IPlayerBusiness playerBusiness, IImageBusiness imageBusiness)
         {
             _context = context;
+            _teamBusiness = teamBusiness;
+            _cryption = cryption;
+            _playerBusiness = playerBusiness;
+            _imageBusiness = imageBusiness;
         }
 
 
         /// <summary>
         /// Method to Add the Player passed in the parameters to the database
         /// </summary>
-        /// <param name="player">Object Player to Add</param>
-        public void Add(Player player)
+        /// <param name="game">Object Player to Add</param>
+        public async Task<ApiResponse> Add(Player player)
         {
-           _context.Players.Add(player);
-           _context.SaveChanges();
+            try
+            {
+                if (!await _playerBusiness.Check(player))
+                {
+                    player.Picture = await _imageBusiness.UploadImage(player.ImageBase64, $"{player.FirstName}-{player.LastName}-{player.Login}-{DateTime.Now.ToString("yyyyMMddHHmmss")}");
+                    player.Password = _cryption.Encrypt(player.Password);
+                    await _context.Players.AddAsync(player);
+                    await _context.SaveChangesAsync();
+                    return new ApiResponse { Status = ApiStatus.Ok, Message = ApiAction.Add, Response = player };
+                } else
+                {
+                    return new ApiResponse { Status = ApiStatus.CantDelete, Message = "Le pseudo et l'e-mail doivent être uniques."};
+                }
+            }
+            catch (Exception e)
+            {
+                return TranslateError.Convert(e);
+            }
         }
 
         /// <summary>
         /// Method to Delete from the database the Player passed in the parameters
         /// </summary>
         /// <param name="player">Object Player to Delete</param>
-        public void Delete(Player player)
+        public async Task<ApiResponse> Delete(int idPlayer)
         {
-           _context.Players.Remove(player);
-           _context.SaveChanges();
+            try
+            {
+                if (!await _teamBusiness.OnlyThisPlayer(idPlayer))
+                {
+                    if (!await _teamBusiness.ContainsTeamsWithPlayer(idPlayer))
+                    {
+                       
+                        await _teamBusiness.DeleteAllFromPlayer(idPlayer);
+                        var player = await _context.Players.FindAsync(idPlayer);
+                        if (player.Picture != null)
+                        {
+                            _imageBusiness.DeleteImage(player.Picture);
+                        }
+                        _context.Players.Remove(player);
+                        await _context.SaveChangesAsync();
+                        return new ApiResponse { Status = ApiStatus.Ok, Message = ApiAction.Delete };
+                    }
+                    else
+                    {
+                        return new ApiResponse { Status = ApiStatus.CantDelete, Message = "Ce participant participe ou a déjà participé à un jeu !" };
+                    }
+
+                }
+                else
+                {
+                    return new ApiResponse { Status = ApiStatus.CantDelete, Message = "Impossible de supprimer ce participant, il est dans une équipe seul !" };
+                }
+            }
+            catch (Exception e)
+            {
+                return TranslateError.Convert(e);
+            }
+
         }
 
         /// <summary>
@@ -71,10 +137,37 @@ namespace DijiWalk.Repositories
         /// Method that will Update the Player passed in the parameters to the database
         /// </summary>
         /// <param name="player">Object Player to Update</param>
-        public void Update(Player player)
+        public async Task<ApiResponse> Update(Player player)
         {
-           _context.Players.Update(player);
-           _context.SaveChanges();
+            try
+            {
+                if (!await _playerBusiness.CheckUpdate(player))
+                {
+                    if (player.ImageChanged)
+                    {
+                        if (player.Picture != null)
+                        {
+                            _imageBusiness.DeleteImage(player.Picture);
+                        }
+                        player.Picture = await _imageBusiness.UploadImage(player.ImageBase64, $"{player.FirstName}-{player.LastName}-{player.Login}-{DateTime.Now.ToString("yyyyMMddHHmmss")}");
+                    }
+                    if (player.PasswordChanged)
+                    {
+                        player.Password = _cryption.Encrypt(player.Password);
+                    }   
+                    _context.Players.Update(player);
+                    await _context.SaveChangesAsync();
+                    return new ApiResponse { Status = ApiStatus.Ok, Message = ApiAction.Update, Response = player };
+                }
+                else
+                {
+                    return new ApiResponse { Status = ApiStatus.CantDelete, Message = "Le pseudo et l'e-mail doivent être uniques." };
+                }
+            }
+            catch (Exception e)
+            {
+                return TranslateError.Convert(e);
+            }
         }
     }
 }
